firmware: multi-target HAL, config v5, logic engine, presence fusion, Web UI settings
- HAL dispatcher + per-target profiles (svc_hal, hal_svc100, hal_kincony) — SVC-021/023
- config schema v5 + v1..v4->v5 migration + board-id guard — SVC-026
- logic engine v0.1 + control_logic adapter (gated) — SVC-013
- 2-sensor presence fusion (pure, host-tested)
- Web UI settings: POST /api/config, whitelist apply, auth+CSRF, bounded/escaped
  JSON, constant-time token compare — SVC-014
- host tests (ASan/UBSan): ALL PASS

+++ components/dinput/include/dinput.h
@@ -20,7 +20,7 @@ extern "C" {
 /** @brief Driver configuration. */
 typedef struct {
     uint16_t debounce_ms;    /**< Stable window before a change is accepted.   */
-    uint8_t  active_low;     /**< Bitmask: bit i set => channel i is active-low.*/
+    uint32_t active_low;     /**< Bitmask: bit i set => channel i is active-low.*/
 } dinput_cfg_t;
 
 /**
@@ -44,7 +44,7 @@ svc_err_t dinput_start(const dinput_cfg_t *cfg);
 svc_err_t dinput_get(uint8_t channel, bool *out_active);
 
 /** @brief Bitmask snapshot of debounced logical-active states. */
-uint8_t dinput_state_mask(void);
+uint32_t dinput_state_mask(void);
 
 #ifdef __cplusplus
 }
